1. Mapster32 saves autosave variable in the cfg. 2. Autosaving interval is configurable. 3. Sprite noclip is on by default(CTRL+N to toggle). 4. Fixed the devastator crash. Thanks to Nukey for helping me to track it down.
Notes about the devastator crash.
The yvel of projectile may point to an already removed sprite(its sectnum is MAXSECT). If some code tries to access the sector of the removed sprite, the crash might happen.

How to reproduce crashes.
True story: Nukey recorded a demo(*.dmo) and send me his demo and map so that I could easily reproduce it and eventually fix the crash.

git-svn-id: http://svn.eduke32.com/eduke32@744 1a8010ca-5511-0410-912e-c29ae57300e0