Replace the semicolon trailing empty-bodied loops with opening and closing braces.
I believe this is the clearest way to indicate that an empty-bodied loop is intended, as opposed to the possibility of a mistaken semicolon.

This eliminates "warning: suggest a space before ‘;’ or explicit braces around empty body in ‘for’ statement" with GCC 4.2, which is important for old OS X versions.

git-svn-id: http://svn.eduke32.com/eduke32@5556 1a8010ca-5511-0410-912e-c29ae57300e0